test(yamlgen): give Tagged.Decoded the untagged Go type for "!!int"
The library decodes a "!!int" to the Go type the untagged number takes --
uint64, int64 or *big.Int -- where it gave a plain int, and a merge's own
key beats a merged time.Time key at the same instant in another zone. The
generator follows both.

- Tagged.Decoded and readings.legacy drop the arm that made "!!int" an int.
- keyedMap.holds finds a time.Time key at the same instant, as codec's
  keyedMap does for a merge; the map keeps the own key's zone.
- Five hand-written tests expect uint64 where they expected int.
- The Timestamp doc no longer claims a plain "2001-12-14" is a string under
  every version: YAML 1.1 reads it as a timestamp, and so does the library
  under a "%YAML 1.1" directive. The emitter still quotes every drawn date.

+++ internal/testintegration/yamlcorpus/key_naming_test.go
@@ -290,7 +290,7 @@ func TestATagIsNoLaxerThanTheBareSpelling(t *testing.T) {
 	})
 
 	t.Run("and 1.1's sexagesimal reaches both tags", func(t *testing.T) {
-		assert.Equal(t, 685230, valueOf(t, under11+"k: !!int 190:20:30\n"))
+		assert.Equal(t, uint64(685230), valueOf(t, under11+"k: !!int 190:20:30\n"))
 		assert.Equal(t, 685230.5, valueOf(t, under11+"k: !!float 190:20:30.5\n"))
 		// Neither is a number under the core schema.
 		assert.Contains(t, refuses(t, "k: !!int 190:20:30\n"), "as !!int")
